Refactoring: Improve SchemaProcessor.extractName() to also return the namespaceUri (#66)
This is a pure refactoring (i.e. no functional change) because the
result object of extractName() is never used as a whole.  Instead, the
caller functions only use the specific fields they are interested in,
and ignore all other fields.

In other words, any new fields in the result of extractName() are
ignored by the rest of the code base.